+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## Unreleased
+- dev: check plugin dependency versions.
 - dev: namespace Composer dependencies with Strauss.
-- chore: update WPGraphQL Plugin Boilerplate dependency to v0.0.8.
 - dev: wrap global functions in `function_exists()` checks.
+- chore: update WPGraphQL Plugin Boilerplate dependency to v0.0.8.
+
 ## v0.0.9
 - chore: update WPGraphQL Plugin Boilerplate dependency to v0.0.7.